+++ NEWS.md
@@ -1,3 +1,5 @@
+# pkgcache (development version)
+
 # pkgcache 2.0.2
 
 * pkgcache error messages are better now.